requirements.txt
streamlit
google-generativeai
librosa
numpy
scipy
Pillow
matplotlib